+++ logic/src/main/java/org/itsallcode/whiterabbit/logic/Config.java
@@ -19,6 +19,8 @@ public interface Config
 
     Optional<Duration> getCurrentHoursPerDay();
 
+    Optional<Duration> getMandatoryBreak();
+
     default Path getProjectFile()
     {
         return getDataDir().resolve(PROJECTS_JSON);